@@ -11,7 +11,7 @@ describe XPath::Union do
       @expr1 = XPath.generate { |x| x.descendant(:p) }
       @expr2 = XPath.generate { |x| x.descendant(:div) }
       @collection = XPath::Union.new(@expr1, @expr2)
-      @collection.expressions.should eq [@expr1, @expr2]
+      expect(@collection.expressions).to eq [@expr1, @expr2]
     end
   end
 
@@ -22,7 +22,7 @@ describe XPath::Union do
       @collection = XPath::Union.new(@expr1, @expr2)
       exprs = []
       @collection.each { |expr| exprs << expr }
-      exprs.should eq [@expr1, @expr2]
+      expect(exprs).to eq [@expr1, @expr2]
     end
   end
 
@@ -31,7 +31,7 @@ describe XPath::Union do
       @expr1 = XPath.generate { |x| x.descendant(:p) }
       @expr2 = XPath.generate { |x| x.descendant(:div) }
       @collection = XPath::Union.new(@expr1, @expr2)
-      @collection.map(&:expression).should eq %i[descendant descendant]
+      expect(@collection.map(&:expression)).to eq %i[descendant descendant]
     end
   end
 
@@ -41,9 +41,9 @@ describe XPath::Union do
       @expr2 = XPath.generate { |x| x.descendant(:div).where(x.attr(:id) == 'foo') }
       @collection = XPath::Union.new(@expr1, @expr2)
       @results = doc.xpath(@collection.to_xpath)
-      @results[0][:title].should eq 'fooDiv'
-      @results[1].text.should eq 'Blah'
-      @results[2].text.should eq 'Bax'
+      expect(@results[0][:title]).to eq 'fooDiv'
+      expect(@results[1].text).to eq 'Blah'
+      expect(@results[2].text).to eq 'Bax'
     end
   end
 
@@ -55,9 +55,9 @@ describe XPath::Union do
       @xpath1 = @collection.where(XPath.attr(:id) == 'foo').to_xpath
       @xpath2 = @collection.where(XPath.attr(:id) == 'fooDiv').to_xpath
       @results = doc.xpath(@xpath1)
-      @results[0][:title].should eq 'fooDiv'
+      expect(@results[0][:title]).to eq 'fooDiv'
       @results = doc.xpath(@xpath2)
-      @results[0][:id].should eq 'fooDiv'
+      expect(@results[0][:id]).to eq 'fooDiv'
     end
   end
 end
