@@ -19,7 +19,7 @@ module XPath
       arguments.each(&block)
     end
 
-    def method_missing(*args) # rubocop:disable Style/MethodMissingSuper, Style/MissingRespondToMissing
+    def method_missing(*args) # rubocop:disable Style/MissingRespondToMissing
       XPath::Union.new(*arguments.map { |e| e.send(*args) })
     end
 
